services/trainer/airflow_dags/train_recommendation_model.py
"""
Airflow DAG for training recommendation models.
Schedules periodic model training and evaluation.
"""

from datetime import datetime, timedelta
from airflow import DAG
from airflow.operators.python import PythonOperator
from airflow.operators.bash import BashOperator

# Default arguments
default_args = {
    'owner': 'ml-team',
    'depends_on_past': False,
    'email_on_failure': True,
    'email_on_retry': False,
    'retries': 2,
    'retry_delay': timedelta(minutes=5),
}

# DAG definition
dag = DAG(
    'train_recommendation_model',
    default_args=default_args,
    description='Train and evaluate recommendation model',
    schedule_interval=timedelta(days=1),  # Run daily
    start_date=datetime(2024, 1, 1),
    catchup=False,
    tags=['ml', 'recommendation'],
)

# Task 1: Train model
train_task = PythonOperator(
    task_id='train_model',
    python_callable=lambda: __import__('subprocess').run(
        ['python', 'services/trainer/train.py'],
        check=True
    ),
    dag=dag,
)

# Task 2: Evaluate model (placeholder - add evaluation logic)
evaluate_task = BashOperator(
    task_id='evaluate_model',
    bash_command='echo "Model evaluation completed"',
    dag=dag,
)

# Task 3: Promote model to staging (if metrics are good)
promote_task = BashOperator(
    task_id='promote_to_staging',
    bash_command='echo "Model promoted to staging"',
    dag=dag,
)

# Task dependencies
train_task >> evaluate_task >> promote_task

