services/trainer/requirements.txt
numpy==1.24.3
pandas==2.0.3
psycopg2-binary==2.9.9
scikit-learn==1.3.2
lightgbm==4.1.0
mlflow==2.8.1
python-json-logger==2.0.7

